Deep Learning Project/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=20.1.0=pypi_0
astroid=2.4.2=pypi_0
atari-py=0.2.6=pypi_0
attrs=20.1.0=pypi_0
backcall=0.2.0=pypi_0
backports-ssl-match-hostname=3.7.0.1=pypi_0
blas=1.0=mkl
bleach=3.1.5=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2020.10.14=0
captum=0.3.0=dev_0
certifi=2020.11.8=py37haa95532_0
cffi=1.14.2=pypi_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
colorama=0.4.3=pypi_0
cpuonly=1.0=0
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
cytoolz=0.11.0=py37he774522_0
dask-core=2.27.0=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
fasttext=0.9.2=pypi_0
freetype=2.10.2=hd328e21_0
future=0.18.2=py37_1
gym=0.17.3=pypi_0
h5py=2.10.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pypi_0
imageio=2.9.0=py_0
importlib-metadata=1.7.0=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=pypi_0
ipython=7.18.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
isort=5.5.3=pypi_0
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.17.0=py_0
jpeg=9b=hb83a4c4_2
json5=0.9.5=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab=2.2.6=pypi_0
jupyterlab-server=1.2.0=pypi_0
kiwisolver=1.2.0=py37h74a9793_0
lazy-object-proxy=1.4.3=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_1
markupsafe=1.1.1=pypi_0
matplotlib=3.1.1=pypi_0
matplotlib-base=3.2.2=py37h64f37c6_0
mccabe=0.6.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.1.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mock=4.0.2=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.7=pypi_0
networkx=2.5=py_0
ninja=1.10.1=py37h7ef1ec2_0
nltk=3.5=pypi_0
nose=1.3.7=pypi_0
notebook=6.1.3=pypi_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
olefile=0.46=py37_0
opencv-python=4.4.0.44=pypi_0
openssl=1.1.1h=he774522_0
packaging=20.4=pypi_0
pandas=1.1.3=py37ha925a31_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.2.0=py37hcc1f983_0
pip=20.2.2=py37_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.7=pypi_0
pybind11=2.6.1=pypi_0
pycparser=2.20=pypi_0
pyglet=1.5.0=pypi_0
pygments=2.6.1=pypi_0
pylint=2.6.0=pypi_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=pypi_0
pyrsistent=0.16.0=pypi_0
python=3.7.1=h8c8aaf0_6
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.7_cpu_0
pytz=2020.1=py_0
pywavelets=1.1.1=py37he774522_0
pywin32=228=pypi_0
pywinpty=0.5.7=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=pypi_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.7=pypi_0
qtpy=1.9.0=pypi_0
regex=2020.10.11=pypi_0
requests=2.24.0=pypi_0
scikit-image=0.16.2=py37h47e9c7a_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=pypi_0
send2trash=1.5.0=pypi_0
setuptools=49.6.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.11.0=pypi_0
sqlite=3.33.0=h2a8f88b_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.1=pypi_0
toolz=0.11.1=py_0
torchvision=0.7.0=py37_cpu
tornado=6.0.4=py37he774522_1
tqdm=4.50.2=pypi_0
traitlets=4.3.3=pypi_0
typed-ast=1.4.1=pypi_0
urllib3=1.25.10=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
